+++ crates/config/tests/generator_test.rs
@@ -52,6 +52,9 @@ struct GenConfig {
     spec_unresolved: version_spec::UnresolvedVersionSpec,
     id: warpgate::Id,
     locator: Option<warpgate::PluginLocator>,
+    json_value: serde_json::Value,
+    toml_value: Option<toml::Value>,
+    yaml_value: serde_yaml::Value,
 }
 
 fn create_generator() -> SchemaGenerator {